+++ lstore/db.cpp
@@ -1,20 +1,8 @@
 #include <map>
 #include <string>
 #include <stdexcept>
+#include "db.h"
 #include "table.h"
-
-class Database{
-  public:
-    Database(){};
-    ~Database(){};
-    //void open(path); for next Milestone
-    //void close(); for next Milestone
-    Table create_table(std::string name, int num_columns, int key_index);
-    void drop_table(std::string name);
-    Table get_table(std::string name);
-  private:
-    std::map<std::string, Table> tables;
-};
 
 //Creates a new table
 Table Database::create_table(std::string name, int num_columns, int key_index){
